Fixed the cyclic dependecy issue in the FontSizeOptions component.
