+++ src/components/Editor/Menu/Fixed/constants.js
@@ -1,5 +1,3 @@
-import { t } from "i18next";
-
 import MenuButton from "./components/Button";
 import EmojiOption from "./components/EmojiOption";
 import FontSizeOption from "./components/FontSizeOption";
@@ -7,15 +5,6 @@ import TableOption from "./components/TableOption";
 import TextColorOption from "./components/TextColorOption";
 
 import Mentions from "../../CustomExtensions/Mention";
-
-export const FONT_SIZE_OPTIONS = [
-  { label: t("neetoEditor.menu.heading1"), value: 1, key: "h1" },
-  { label: t("neetoEditor.menu.heading2"), value: 2, key: "h2" },
-  { label: t("neetoEditor.menu.heading3"), value: 3, key: "h3" },
-  { label: t("neetoEditor.menu.heading4"), value: 4, key: "h4" },
-  { label: t("neetoEditor.menu.heading5"), value: 5, key: "h5" },
-  { label: t("neetoEditor.menu.normalText"), value: 0, key: "body2" },
-];
 
 export const MENU_ELEMENT_TYPES = {
   BUTTON: "button",
